add preliminary support for mouse movements
Idea is to use the tobii5 on windows with its "experience" package to
set a coarse position, and "tickle" (jiggle isn't recognized by w2l)
the cursor to make it warp to the right area.  Then on it's either less
movement with the real mouse, or maybe just dictate the fine movement.

Later on zoomed-in mouse-grid could work as well.

Even later on Talon will likely provide a more accurate tobii5 reading
and a nicer zoom mouse, making this thing obsolete.